refactor(hindsight): rename intent.rs to maker.rs
The module finds the order maker's trade; "intent" named the order
style, not what the code does. Matches the TraderRole::Maker vocabulary.
